вносит изменения в форму создания/редактирования - устанавливает значения по умолчанию; переносит константы в отдельный файл, общие функции пемещает с отдельный файл
